requirements.txt
djangorestframework
pytest
pytest-django
Django